+++ src/tasks/mocha/defaults.js
@@ -7,6 +7,7 @@ module.exports = {
   files: path.join(processCwd, 'src/+(test|tests)/**/*.+(js|ts|tsx)'),
   require: `${path.join(__dirname, 'setup-file.js')}`,
   maxConcurrency: 0,
+  runnerMemoryUsageLimit: 0,
   collectCoverage: true,
   coverageReporters: ['text-summary', 'html'],
   coverageDirectory: 'coverage',